Cleaned up CountyIterable
