+++ src/client/counties/CountyIterable.ts
@@ -1,31 +1,10 @@
 import { Page } from 'puppeteer';
 import { County } from '../types/County';
 import { State } from '../types/State';
+import parseCounty from './utils/parseCounty';
+import parseState from './utils/parseState';
 
-const ORIGIN = 'https://mugshots.com';
-
-const toCounty = (path: string): County | null => {
-  const split = path.split('/');
-
-  if (split.length < 4) return null;
-
-  return {
-    state: format(split[2]),
-    name: format(split[3]),
-    url: ORIGIN + path
-  };
-};
-
-const toState = (path: string): State => {
-  return {
-    name: path.split('/')[2],
-    url: ORIGIN + path
-  };
-};
-
-const format = (str: string): string => {
-  return str.replace(/-/g, ' ');
-};
+const origin = 'https://mugshots.com';
 
 const scrapeCountyHrefs = (page: Page): Promise<string[]> => {
   return page.evaluate(() => {
@@ -42,17 +21,17 @@ const scrapeStateHrefs = (page: Page): Promise<string[]> => {
 };
 
 const getStates = async (page: Page): Promise<State[]> => {
-  await page.goto(`${ORIGIN}/US-States/`);
+  await page.goto(`${origin}/US-States/`);
   const hrefs = await scrapeStateHrefs(page);
-  return hrefs.map(toState).filter(state => state.name);
+  return hrefs.map(parseState).filter(state => state.name);
 };
 
 const getCounties = async (page: Page, state: State): Promise<County[]> => {
   await page.goto(state.url);
   const hrefs = await scrapeCountyHrefs(page);
 
-  return hrefs.reduce((results: County[], path: string) => {
-    const county = toCounty(path);
+  return hrefs.reduce((results: County[], href: string) => {
+    const county = parseCounty({ href, origin });
     if (county) results.push(county);
     return results;
   }, []);
