+++ src/client/counties/utils/parseState.ts
@@ -1,0 +1,10 @@
+import { State } from '../../types/State';
+
+const parseState = (path: string): State => {
+  return {
+    name: path.split('/')[2],
+    url: origin + path
+  };
+};
+
+export default parseState;
